requirements-dev.txt
eth-brownie>=1.14.6,<2.0.0
